backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
opencv-python==4.8.1.78
numpy==1.24.3  # Downgraded for OpenCV compatibility
pillow==10.1.0
python-multipart==0.0.6
websockets==12.0
pydantic==2.5.0
reportlab==4.0.7
python-dotenv==1.0.0